+++ composer/algorithms/mixup/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.mixup.mixup import MixUp as MixUp
-from composer.algorithms.mixup.mixup import MixUpHparams as MixUpHparams
 from composer.algorithms.mixup.mixup import gen_mixup_interpolation_lambda as gen_mixup_interpolation_lambda
 from composer.algorithms.mixup.mixup import mixup_batch as mixup_batch
